Refactor add new files for request and response classes
