+++ digiprod_gen/backend/io/io_fns.py
@@ -5,7 +5,7 @@ import streamlit as st
 
 from io import BytesIO
 from PIL import Image
-from digiprod_gen.backend.models.mba import CrawlingMBARequest
+from digiprod_gen.backend.models.request import CrawlingMBARequest
 
 
 @st.cache_data
