+++ digiprod_gen/backend/models/response.py
@@ -1,0 +1,8 @@
+from typing import List
+
+from pydantic import BaseModel
+
+
+class UploadMBAResponse(BaseModel):
+    warnings: List[str]
+    errors: List[str]
