+++ digiprod_gen/frontend/tab/upload/listing_generation.py
@@ -6,11 +6,11 @@ from langchain.chat_models import ChatOpenAI
 
 from digiprod_gen.backend.text.text_gen_fns import get_product_text_gen, remove_banned_words
 from digiprod_gen.backend.prompt_engineering.utils import extract_list_from_output
-from digiprod_gen.frontend.session import read_session
-
-from digiprod_gen.backend.models.mba import CrawlingMBARequest, MBAProductTextType, MBAMarketplaceDomain
+from digiprod_gen.backend.models.mba import MBAProductTextType, MBAMarketplaceDomain
+from digiprod_gen.backend.models.request import CrawlingMBARequest
 from digiprod_gen.backend.models.session import SessionState
 from digiprod_gen.backend.text.mba_banned_word import MBA_BANNED_WORDS
+from digiprod_gen.frontend.session import read_session
 
 
 def listing_generation(st_tab_ig: DeltaGenerator):
